Fixed a problem when -o is used and not -n.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/kadeploy3/trunk@209 4ab4880d-aebc-4107-8cb6-a8fee4233f52

+++ src/kareboot.rb
@@ -271,6 +271,8 @@ if (exec_specific_config != nil) then
         system("cat #{file.path} >> #{exec_specific_config.nodes_ok_file}")
       }
     end
+  end
+  if (exec_specific_config.nodes_ko_file != "") then
     File.delete(exec_specific_config.nodes_ko_file) if File.exist?(exec_specific_config.nodes_ko_file)
     if (not files_ko_nodes.empty?) then
       files_ko_nodes.each { |file|
